zfs: move Linux-specific code out of zfs.go
Avoid Linux-specific code scattered in two places by moving it to the
already-existing zfs_linux.go.
